refactor(auth): user, auth 도메인 분리. 패키지 구조 변경
