+++ src/main/java/doritos/doriroom/global/jwt/JwtUtil.java
@@ -1,22 +1,18 @@
 package doritos.doriroom.global.jwt;
 
-import doritos.doriroom.global.exception.ApiException;
-import doritos.doriroom.user.domain.RefreshToken;
+import doritos.doriroom.auth.domain.RefreshToken;
 import doritos.doriroom.user.exception.UsernameNotFoundException;
-import doritos.doriroom.user.repository.RefreshTokenRedisRepository;
+import doritos.doriroom.auth.repository.RefreshTokenRedisRepository;
 import doritos.doriroom.user.domain.User;
 import doritos.doriroom.user.repository.UserRepository;
 import io.jsonwebtoken.security.Keys;
-import jakarta.servlet.ServletException;
 import lombok.*;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import io.jsonwebtoken.*;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
 import java.util.Date;
-import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
